chore: Update ACL rules for Examiner and Invigilator roles
This commit updates the ACL rules in the `configs/acl.js` file to include the 'Exam' subject for the 'EXAMINER' role and the 'ExamPaper' and 'Exam' subjects for the 'INVIGILATOR' role. This ensures that the appropriate permissions are granted to these roles for managing and reading exams and exam papers.

Implemented UI for exam, exam-papers and audit